refactor: rename HandRanking to HandRank and move to separate file
