+++ src/main/kotlin/hwr/oop/projects/peakpoker/core/hand/HandRank.kt
@@ -1,0 +1,14 @@
+package hwr.oop.projects.peakpoker.core.hand
+
+enum class HandRank {
+    HIGH_CARD,
+    ONE_PAIR,
+    TWO_PAIR,
+    THREE_OF_A_KIND,
+    STRAIGHT,
+    FLUSH,
+    FULL_HOUSE,
+    FOUR_OF_A_KIND,
+    STRAIGHT_FLUSH,
+    ROYAL_FLUSH
+}
